Compile template_loop as its own compilation unit
